refactor(code): split code into smaller modules
