+++ .env.example
@@ -1,4 +1,2 @@
 # you can get your gemini api key from https://aistudio.google.com/app/apikey
 API_KEY=your_gemini_api_key_here
-
-VERCEL_URL="localhost:3000"
