+++ action/index.ts
@@ -1,67 +1,35 @@
 "use server";
 
-import { model } from "@/utils";
+import { handleModelResponse } from "@/utils";
+import { generateSystemInstruction } from "@/utils/instructions";
 
-export const commitChange = async ({
-  message,
-  isEmojiSupport,
-}: {
+type ActionProps = {
   message: string | null;
   isEmojiSupport: boolean;
-}): Promise<{
+};
+
+type ActionPromise = {
   data: {
     text: string;
   } | null;
   error: string | null;
-}> => {
+};
+
+export const commitChange = async ({
+  message,
+  isEmojiSupport,
+}: ActionProps): Promise<ActionPromise> => {
   if (!message || message.trim() === "") {
     return { data: null, error: "Please enter a message" };
   }
 
   try {
-    const modelResponse = model.generateContent({
-      contents: [{ role: "user", parts: [{ text: message }] }],
-      systemInstruction: `\
-      You are an assistant that helps to provide Git commit messages based on https://www.conventionalcommits.org/en/v1.0.0/.
-
-      - Provide a Git commit message in a code block as txt.
-      - Do not include the full command (e.g., \`git commit -m "here git message"\`).
-      - Only provide the commit message itself.
-      - Suggest 3 different commit messages to give the user some options.
-      - For example, if the user input is "I change lib folder to utils folder", then the output should be:
-
-      ${
-        isEmojiSupport &&
-        `\
-      - Use emojis in the commit message.
-      - For emojis, you can use https://gitmoji.dev/
-      - Don't provide a description, just the commit message.
-      - For example, if the user input is "I change lib folder to utils folder", then the output should be:
-      `
-      }
-
-      ${
-        isEmojiSupport
-          ? `
-      \`\`\`txt \n ♻️ refactor(lib): change lib folder to utils folder \n\`\`\`\n
-      \`\`\`txt \n ➕ refactor(deps): rename lib folder to utils \n\`\`\`\n
-      \`\`\`txt \n ✏️ fix(deps): rename lib folder to utils \n\`\`\`\n
-          `
-          : `
-      \`\`\`txt \n refactor(lib): change lib folder to utils folder \n\`\`\`\n
-      \`\`\`txt \n refactor(deps): rename lib folder to utils \n\`\`\`\n
-      \`\`\`txt \n fix(deps): rename lib folder to utils \n\`\`\`\n
-      `
-      }
-
-      `,
-    });
-
-    const response = (await modelResponse).response.text();
+    const systemInstruction = generateSystemInstruction(isEmojiSupport);
+    const responseText = await handleModelResponse(systemInstruction, message);
 
     console.log({
       message: message,
-      response: response,
+      response: responseText,
       data: {
         time: new Date().toISOString(),
       },
@@ -69,12 +37,15 @@ export const commitChange = async ({
 
     return {
       data: {
-        text: response,
+        text: responseText,
       },
       error: null,
     };
-  } catch (error) {
+  } catch (error: any) {
     console.log("error on action", error);
-    return { data: null, error: "something went wrong!" };
+    return {
+      data: null,
+      error: error?.statusText ? error.statusText : "something went wrong!",
+    };
   }
 };
