+++ utils/index.ts
@@ -1,7 +1,5 @@
 import { GoogleGenerativeAI } from "@google/generative-ai";
 
-export const runtime = "node";
-
 // Get your API key from https://makersuite.google.com/app/apikey
 if (!process.env.API_KEY) {
   console.log("API_KEY environment variable is required", process.env.NODE_ENV);
@@ -10,6 +8,16 @@ if (!process.env.API_KEY) {
 
 const genAI = new GoogleGenerativeAI(process.env.API_KEY!);
 
-export const model = genAI.getGenerativeModel({
+const model = genAI.getGenerativeModel({
   model: "gemini-1.5-pro-latest",
 });
+
+export const handleModelResponse = async (systemInstruction: string, message: string) => {
+  const modelResponse = model.generateContent({
+    contents: [{ role: "user", parts: [{ text: message }] }],
+    systemInstruction,
+  });
+
+  const response = (await modelResponse).response.text();
+  return response;
+};
