+++ utils/instructions.ts
@@ -1,0 +1,31 @@
+export const generateSystemInstruction = (
+  isEmojiSupport: boolean
+) => {
+  const baseInstruction = `\
+  You are an assistant that helps to provide Git commit messages based on https://www.conventionalcommits.org/en/v1.0.0/.
+  
+  - Provide a Git commit message in a code block as txt.
+  - Do not include the full command (e.g., \`git commit -m "here git message"\`).
+  - Only provide the commit message itself.
+  - Suggest 3 different commit messages to give the user some options.
+  - For example, if the user input is "I change lib folder to utils folder", then the output should be:
+  `;
+
+  const emojiInstruction = isEmojiSupport
+    ? `
+  - Use emojis in the commit message.
+  - For emojis, you can use https://gitmoji.dev/
+  - Don't provide a description, just the commit message.
+  - For example, if the user input is "I change lib folder to utils folder", then the output should be:
+  \`\`\`txt \n ♻️ refactor(lib): change lib folder to utils folder \n\`\`\`\n
+  \`\`\`txt \n ➕ refactor(deps): rename lib folder to utils \n\`\`\`\n
+  \`\`\`txt \n ✏️ fix(deps): rename lib folder to utils \n\`\`\`\n
+  `
+    : `
+  \`\`\`txt \n refactor(lib): change lib folder to utils folder \n\`\`\`\n
+  \`\`\`txt \n refactor(deps): rename lib folder to utils \n\`\`\`\n
+  \`\`\`txt \n fix(deps): rename lib folder to utils \n\`\`\`\n
+  `;
+
+  return baseInstruction + emojiInstruction;
+};
